refactor : 프로젝트 구조 및 test 구조 수정
