+++ src/main/java/com/team/comma/spotify/recommend/domain/Recommend.java
@@ -31,6 +31,7 @@ public class Recommend {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     @Enumerated(EnumType.STRING)
     private RecommendType recommendType;
