+++ src/main/java/com/team/comma/util/security/domain/RefreshToken.java
@@ -22,7 +22,6 @@ public class RefreshToken {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(nullable = false)
     private Long id;
 
     @Column(nullable = false)
