+++ src/test/java/com/team/comma/util/security/controller/SecurityControllerTest.java
@@ -1,11 +1,31 @@
-package com.team.comma.security;
+package com.team.comma.util.security.controller;
+
+import static com.team.comma.common.constant.ResponseCode.ACCESS_TOKEN_CREATE;
+import static com.team.comma.common.constant.ResponseCode.AUTHORIZATION_ERROR;
+import static com.team.comma.common.constant.ResponseCode.LOGOUT_SUCCESS;
+import static org.apache.http.cookie.SM.SET_COOKIE;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
+import static org.springframework.restdocs.cookies.CookieDocumentation.cookieWithName;
+import static org.springframework.restdocs.cookies.CookieDocumentation.requestCookies;
+import static org.springframework.restdocs.cookies.CookieDocumentation.responseCookies;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessRequest;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
+import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
+import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.google.gson.Gson;
-import com.team.comma.security.oauth.SecurityController;
-import com.team.comma.confused.dto.MessageResponse;
-import com.team.comma.exception.FalsifyTokenException;
-import com.team.comma.security.jwt.service.JwtService;
+import com.team.comma.common.dto.MessageResponse;
+import com.team.comma.util.jwt.exception.FalsifyTokenException;
+import com.team.comma.util.jwt.service.JwtService;
+import com.team.comma.util.security.controller.SecurityController;
 import jakarta.servlet.http.Cookie;
+import java.nio.charset.StandardCharsets;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -26,19 +46,6 @@ import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
-
-import java.nio.charset.StandardCharsets;
-
-import static com.team.comma.constant.ResponseCode.*;
-import static org.apache.http.cookie.SM.SET_COOKIE;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.*;
-import static org.springframework.restdocs.cookies.CookieDocumentation.*;
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
-import static org.springframework.restdocs.payload.PayloadDocumentation.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @AutoConfigureRestDocs
 @ExtendWith({RestDocumentationExtension.class, SpringExtension.class})
