+++ src/test/java/com/team/comma/util/security/repository/RefreshTokenRepositoryTest.java
@@ -1,15 +1,15 @@
-package com.team.comma.confused;
+package com.team.comma.util.security.repository;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
 
-import com.team.comma.confused.security.RefreshTokenRepository;
+import com.team.comma.util.security.domain.RefreshToken;
+import com.team.comma.util.security.repository.RefreshTokenRepository;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-
-import com.team.comma.confused.security.RefreshToken;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
